feat(contact/page.tsx): add contact page component feat(experience/page.tsx): add experience page component
The contact page component is added to the application, allowing users to fill out a form and send a message. The form data is sent using the emailjs library. Success and error messages are displayed using the react-toastify library.

The experience page component is added to the application, displaying a timeline of work experiences. The experiences are rendered using the ExperienceCard component and styled using the react-vertical-timeline-component library.